structuring modified further
