--- problem1/requirement/logic_test.go
@@ -1,1 +1,0 @@
-package requirement
