+++ problem1/util/util.go
@@ -1,4 +1,4 @@
-package concrete
+package util
 
 import "fmt"
